requirements.txt
pycryptodome>=3.19.0
numpy>=1.24.0
pytest>=7.0.0
